Refactor/Valores financeiros mudando de Double para BigDecimal

+++ codetech/src/main/java/tech/code/codetech/model/Faturamento.java
@@ -3,6 +3,8 @@ package tech.code.codetech.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -14,7 +16,7 @@ public class Faturamento {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
-    private Double lucro;
+    private BigDecimal lucro;
 
     @OneToOne
     private OrdemServico ordemServico;
